Mostly tidying
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@4465 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSTimeZone.m
@@ -729,7 +729,7 @@ decode (const void *ptr)
 
 	detail = [[NSConcreteTimeZoneDetail alloc]
 			initWithTimeZone: zone
-			      withAbbrev: abbrevsArray[types[i].abbr_idx]
+			      withAbbrev: abbrevsArray[(int)types[i].abbr_idx]
 			      withOffset: types[i].offset
 				 withDST: (types[i].isdst > 0)];
 	[detailsArray addObject: detail];
